Add files via upload
Objective: Publish your static resume created in Milestone 1 to GitHub and deploy it live using Vercel.

Requirements:  Push the project code to a GitHub repository.  Deploy the project on Vercel so that the resume is publicly accessible.

Scoring Breakdown:  GitHub Repository Setup and Organization (50 points): o Is the code properly committed and pushed to a GitHub repository? o Is the repository well-organized with proper folder structures and file naming

conventions?  Successful Deployment on Vercel (50 points): o Is the resume successfully deployed and accessible online via Vercel?